Convert build system to autotools.
A pretty straight-forward conversion, fixes #59. The Documentation
makefile is left alone (we just ship the manpages in there and don't
do any processing at build/install time).

From a fresh clone, do:

./autogen.sh
./configure
make

Change details:

* add local copy of nvme_ioctl.h (from kernel 4.5 rc6), remove local
  ioctl defs from nvme.h. Use autoconf to prefer system header, fall
  back to internal header.

* Rename and update rpm spec file to work with autotools and other
  misc cleanups. Tested on RHEL7.0.

* Debian packaging needs some tweaks to work still, the targets have
  just been copied verbatim to the bottom of Makefile.am.
