+++ autogen.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+mkdir -p config
+autoreconf -ivf
